% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{as_distance_vector}
\alias{as_distance_vector}
\title{Convert Latent Locations to a Vector of Distances}
\usage{
as_distance_vector(Z, ...)
}
\arguments{
\item{Z}{Matrix representing latent locations.}

\item{...}{Additional parameters to be passed to \code{\link{dist}}.}
}
\value{
Column vector of distances as contained in the \code{lower.tri(d)},
where \code{d} is the distance matrix.
}
\description{
Converts a matrix of latent locations to a vector of pairwise distances.
}
\details{
Converts a matrix of latent locations for each of the \code{n} nodes to a
column vector of pairwise distances.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

