% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynnet.R
\name{c.dynnet}
\alias{c.dynnet}
\title{Concatenate dynnet Objects into Single Composite Network}
\usage{
\method{c}{dynnet}(..., recursive = FALSE)
}
\arguments{
\item{...}{dynnet objects to be concatenated.}

\item{recursive}{Logical. If recursive = TRUE, recursively descend through
lists of dynnet objects, combining all their elements into a single
dynnet object.}
}
\value{
dynnet object.
}
\description{
Concatenate dynnet objects into a single composite network.
}
\details{
Currently, no checking is done to verify the homogeneity of the networks
with respect to the number of vertices, directionality, type of ties,
etc. User beware.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

