% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_positions}
\alias{calc_positions}
\title{Calculate Node Postitions}
\usage{
calc_positions(periods, pos, traj, fn = trajectory_linear, ...)
}
\arguments{
\item{periods}{Integer vector of time periods for which to calculate the
position.}

\item{pos}{Matrix of latent locations.}

\item{traj}{Matrix of trajectories.}

\item{fn}{Function taking a time period, latent starting positions, and
trajectories, which returns a matrix of latent position. Default is the
simple linear trajectory model.}

\item{...}{List of parameters to pass to the trajectory function.}
}
\value{
List of latent locations, one for each time period.
}
\description{
Calculate node positions for a specific time period.
}
\details{
Calculate node positions for a specific time period.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

