% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{default_ref_pos}
\alias{default_ref_pos}
\title{Generate Default Reference Locations}
\usage{
default_ref_pos(k)
}
\arguments{
\item{k}{Positive integer indicating the dimensions of the latent social
space.}
}
\value{
An \code{(k+1)}-by-\code{k} numeric matrix of latent locations on
the \code{k}-dimensional unit hypersphere.
}
\description{
Create random default reference node locations on a \code{k}-dimensional
unit hypersphere.
}
\details{
Create random default reference node locations on a \code{k}-dimensional
unit hypersphere. The minimum distance between any two nodes is guaranteed
to be greater than \code{sqrt(2) / scale} where \code{scale = k-1} when
\code{k > 2}. In 2 dimensions, nodes will never be located within the same
quadrant. \code{scale} is used for larger dimensions to greatly reduce
reference position generation time (using \code{sqrt(2)} meant that many
sets of reference positions were being rejected).
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
