% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate.R
\name{dgp_fn}
\alias{dgp_fn}
\title{Network Data Generating Process}
\usage{
dgp_fn(family)
}
\arguments{
\item{family}{String in \code{c("logit", "poisson")} specifying the data
generating process of ties in the network.}
}
\value{
A function taking single argument, a linear predictor.
}
\description{
Produces a function defining the data generating process for ties in the
network.
}
\details{
Produces a function defining the data generating process for ties in the
network.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

