% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DparBlock}
\alias{DparBlock}
\title{Function that computes a symetric D matric from it Block}
\usage{
DparBlock(nD, nq, par)
}
\arguments{
\item{nD}{an integer}

\item{nq}{an integer}

\item{qvector}{a vector}
}
\value{
a symetric positive define matrix
}
\description{
Function that computes a symetric D matric from it Block
}
