% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynnet-package.R
\docType{package}
\name{dynnet-package}
\alias{dynnet}
\alias{dynnet-package}
\title{Latent Space Models for Static and Dynamic Networks}
\description{
\pkg{dynnet} will provide a set of latent space models for static and
dynamic networks. Currently, the package provides the original Euclidean
model of Hoff, Raftery, & Handcock (HRH; 2002) along with a variation of the
model that uses reference nodes as a method of parameter
identification. Extensions to the Projection Model and dynamic networks is
planned for the near future.
}
\details{
\tabular{ll}{
Package: \tab dynnet \cr
Type: \tab Package\cr
Version: \tab 0.0.9001\cr
Date: \tab 2017-01-07\cr
License: \tab GPL-3 \cr
LazyLoad: \tab yes\cr
}
}
\note{
\pkg{dynnet} is developed by Jason W. Morgan.
}
\author{
Jason W. Morgan (\email{jason.w.morgan@gmail.com})
}
\references{
Hoff, Peter D., Adrian E. Raftery, & Mark S. Handcock (2002)
``Latent Space Approaches to Social Network Analysis'', \emph{Journal of the
American Statistical Association}, 97(460), 1090-1098.
}
\keyword{package}

