% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dynnet.R
\name{dynnet_adjacency}
\alias{dynnet_adjacency}
\title{Dynamic Network from a List of Adjacency Networks}
\usage{
dynnet_adjacency(adj, X = NULL, weighted = FALSE, directed = FALSE,
  bipartite = FALSE)
}
\arguments{
\item{adj}{List of adjacency matrixes. Single matrices, representing a
static network, are also supported.}

\item{X}{A \code{data.frame} of nodal covariates. Default is \code{NULL}.}

\item{weighted}{Indicates whether the network is weighted (non-binary). If a
weighted network is provided, but \code{weighted} is not set to \code{TRUE},
all elements of the matrices in \code{adj} will be converted to binary
values, with any value greater than \code{0} equal to \code{1} and all other
set to \code{0}. Default is \code{FALSE}.}

\item{directed}{Indicates whether the network is directed. Default is
\code{FALSE}.}

\item{bipartite}{Indicates whether the network is bipartite. Default is
\code{FALSE}.}
}
\value{
A \code{dynnet} object.
}
\description{
Generate a dynamic network object from a list of adjacency matrices.
}
\details{
Generate a dynamic network object from a list of adjacency matrices.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

