% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynnet.R
\name{edge_density}
\alias{edge_density}
\alias{edge_density.dynnet}
\title{Network Edge Density}
\usage{
edge_density(network)

\method{edge_density}{dynnet}(network)
}
\arguments{
\item{network}{dynnet network object.}
}
\value{
Edge densities
}
\description{
Get the edge density of each graph in the network
}
\details{
Get the edge density of each graph in the network
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

