% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_param.R
\name{enter_param}
\alias{enter_param}
\title{function that returns the parameters object required in the DynNet function.}
\usage{
enter_param(
  structural.model,
  measurement.model,
  Time,
  subject,
  data,
  Tentry = NULL,
  Event = NULL,
  StatusEvent = NULL,
  basehaz = NULL,
  assocT = NULL,
  p.initlev,
  fix.p.initlev = rep(0, length(p.initlev)),
  p.slope,
  fix.p.slope = rep(0, length(p.slope)),
  varcovRE,
  fix.varcovRE = rep(0, length(varcovRE)),
  transitionmatrix,
  fix.transitionmatrix = rep(0, length(transitionmatrix)),
  parab = NULL,
  fix.parab = rep(0, length(parab)),
  var.errors,
  fix.var.errors = rep(0, length(var.errors)),
  transformationY,
  fix.transformationY = rep(0, length(transformationY)),
  baseline1 = NULL,
  fix.baseline1 = rep(0, length(baseline1)),
  p.X1 = NULL,
  fix.p.X1 = rep(0, length(p.X1)),
  p.asso1 = NULL,
  fix.p.asso1 = rep(0, length(p.asso1)),
  p.asso.int1 = NULL,
  fix.p.asso.int1 = rep(0, length(p.asso.int1)),
  baseline2 = NULL,
  fix.baseline2 = rep(0, length(baseline2)),
  p.X2 = NULL,
  fix.p.X2 = rep(0, length(p.X2)),
  p.asso2 = NULL,
  fix.p.asso2 = rep(0, length(p.asso2)),
  p.asso.int2 = NULL,
  fix.p.asso.int2 = rep(0, length(p.asso.int2))
)
}
\arguments{
\item{structural.model}{a list of 7 arguments used to specify the structural model: \describe{
   \item{\code{fixed.LP0}}{a one-sided linear formula object for specifying the 
   fixed effects in the submodel for the baseline level of processes. Note that 
   there is no need to specify a random effect model for the baseline level of 
   processes as we systematically set a process-specific random intercept (with 
   variance fixed to 1 for identifiability purpose). For identifiability purposes, 
   the mean intercepts are fixed to 0 (not estimated).}

   \item{\code{fixed.DeltaLP}}{a two-sided linear formula object for 
   specifying the response outcomes (one the left part of ~ symbol) 
   and the covariates with fixed-effects (on the right part of ~ symbol) 
   in the submodel for change over time of latent processes.}

   \item{\code{random.DeltaLP}}{a one-sided linear formula object 
   for specifying the random effects in the submodel for change 
   over time of latent processes.}

   \item{\code{trans.matrix}}{a one-sided linear formula object 
   for specifying a model for elements of the transition matrix, 
   which captures the temporal influences between latent processes.}

   \item{\code{fixed.survival}}{a one-sided linear formula object 
   for specifying the covariates in the survival sub-model. In competing 
   risks model, the specification for the two events should be separated 
   by the "|" symbol.}

   \item{\code{interactionY.survival}}{a one-sided linear formula 
   object for specifying the covariates in interaction with the 
   dynamics of the latent processes, in the survival sub-model. 
   In competing risks model, the specification for the two events 
   should be separated by the "|" symbol. Only additional terms 
   should be included (No "*" symbol). Covariates in 
   \code{interactionY.survival} should also be included in \code{fixed.survival.}}

   \item{\code{delta.time}}{indicates the discretisation step to be used for latent processes}
   }}

\item{measurement.model}{is a list of arguments detailed below used to specify the measurement model: \describe{

\item{\code{link}}{ indicates the link functions to be used to transform the outcomes. It takes values in "linear" for a linear transformation and "n-type-d" for a I-splines transformation where "n" indicates the number of nodes, "type" (which takes values in "quant", "manual", "equi") indicates where the nodes are placed, and "d" indicates the degree of the I-splines.}

\item{\code{knots}}{ argument indicates if necessary the place of knots (when placed manually with "manual"), default value is NULL}
}}

\item{Time}{indicates the name of the covariate representing the time}

\item{subject}{indicates the name of the covariate representing the grouping structure}

\item{data}{indicates the data frame containing all the variables for estimating the model.}

\item{Tentry}{name of the variable of entry time}

\item{Event}{name of the variable of event time}

\item{StatusEvent}{name of the variable of event status}

\item{basehaz}{type of baseline hazard function}

\item{assocT}{specifies the type of association between the time-to-event(s) and the latent process(es). Values include "r.intercept" for random intercept, "r.slope" for random slope, "r.intercept/slope" for both random intercept and slope, "c.value" for current value.}

\item{p.initlev}{initial values for the fixed effects of the model describing the initial level of the processes}

\item{p.slope}{initial values for the fixed effects of the model describing the slope of the processes.}

\item{varcovRE}{initial values for the lower-triangular matrix in the Cholesky decomposition of the random-effects variance–covariance matrix.}

\item{transitionmatrix}{initial values for the transition matrix}

\item{var.errors}{initial values for the marker-specific measurement error variances.}

\item{transformationY}{initial values for the marker-specific transformation functions (see links in DynNet help)}

\item{baseline1}{initial values for the baseline hazard function in the time-to-event model (in the survival setting) or in the first transition model (in the competing risks setting).}

\item{p.X1}{initial values for the covariate fixed effects (excluding association and latent-process interaction parameters)in the time-to-event model (in the survival setting) or in the first transition model (in the competing risks setting).}

\item{fix.p.X1}{indicator if the parameters \code{p.X1} are fixed (e.g. not estimated).}

\item{p.asso.int1}{initial values for the interactions between covariates and functions of the latent processes in the time-to-event model (in the survival setting) or in the first transition model (in the competing risks setting).}

\item{fix.p.asso.int1}{indicator if the parameters \code{p.asso.int1} are fixed (e.g. not estimated).}

\item{baseline2}{initial values for the baseline hazard function in the second transition model (competing risks setting only). Default to NULL.}

\item{fix.baseline2}{indicator if the parameters \code{baseline2} are fixed (e.g. not estimated).}

\item{p.X2}{initial values for the covariate fixed effects (excluding association and latent-process interaction parameters) in the second transition model (competing risks setting only). Default to NULL.}

\item{fix.p.X2}{indicator if the parameters \code{p.X2} are fixed (e.g. not estimated).}

\item{p.asso2}{initial values for the association parameters in the second transition model (competing risks setting only). Default to NULL.}

\item{fix.p.asso2}{indicator if the parameters \code{p.asso2} are fixed (e.g. not estimated).}

\item{p.asso.int2}{initial values for the interactions between covariates and functions of the latent processes in the second transition model (competing risks setting only). Default to NULL.}

\item{fix.p.asso.int2}{indicator if the parameters \code{p.asso.int2} are fixed (e.g. not estimated).}
}
\value{
A list with the following elements:
\describe{
  \item{paras.ini}{Vector of initial values for all parameters.}
  \item{Fixed.para.index}{Vector containing the indexes of the parameters to be fixed.}
  \item{Fixed.para.values}{Vector of fixed values corresponding to the parameters specified in \code{Fixed.para.index}.}
}
}
\description{
This function creates the vectors of initial parameter values (paras.ini), 
the indices of user-specified fixed parameters (Fixed.para.index), 
and their corresponding fixed values (Fixed.para.values), which are required to 
run the DynNet function.
}
\examples{
### example 1
Delta <- 1
structural.model = list(fixed.LP0 = ~ 1 + C2 | 1 + C2,
                                      fixed.DeltaLP = L2 | L3  ~ 1 | 1 ,
                                      random.DeltaLP = ~ 1|1,
                                      trans.matrix = ~ 1,
                                      delta.time = Delta)
measurement.model = list(link.functions = list(links = c(NULL,NULL),
                                      knots = list(NULL, NULL)))
                                      
Parameters <- enter_param(structural.model = structural.model,
              measurement.model = measurement.model,
              Time = "time",
              subject = "id",
              data = data,
              p.initlev = c(0.000, 0.059, 0.000, 0.163), 
              fix.p.initlev = c(1, 0, 1, 0), 
              p.slope = c(-0.050, -0.153), 
              varcovRE =  c(1.000, 0.000, 0.322, 0.000, 1.000, 0.000, 0.077, 0.139, 0.000, 0.177), 
              fix.varcovRE = c(1, 1, 0, 1, 1, 1, 0, 0, 1, 0),
              transitionmatrix = c(-0.354, 0.114, 0.116, -0.090), 
              var.errors = c(0.287, 0.554), 
              transformationY = c(1.107, 1.889, 0.881, 1.329) )

mod1 <- DynNet(structural.model = structural.model,
              measurement.model = measurement.model,
              parameters = Parameters,
              option = list(nproc = 1, print.info = TRUE,  MCnr = 10, 
                            univarmaxiter = 7, epsa = 1e-5, epsb = 1e-4, epsd = 1e-2),
              Time = "time",
              subject = "id",
              data = data
)

summary(mod1)

 ### example 2
 Delta <- 0.5
 structural.model = list(fixed.LP0 = ~ 1 + C2 | 1 + C2,
                         fixed.DeltaLP = L1 + L2| L3  ~ 1 + time| 1 + time,
                         random.DeltaLP = ~ 1|1,
                         trans.matrix = ~ 1 + bs(x = time, knots =c(2), 
                                                 intercept = F, degree = 2),
                         delta.time = Delta)
measurement.model = list(link.functions = list(links = c(NULL,NULL, NULL),
                                                   knots = list(NULL, NULL, NULL)))
 Parameters <- enter_param(structural.model = structural.model,
                           measurement.model = measurement.model,
                           Time = "time",
                           subject = "id",
                           data = data,
                           p.initlev = c(0.000,0.065, 0.000, 0.168), 
                           fix.p.initlev = c(1, 0, 1, 0), 
                           p.slope = c(-0.054, 0.000, -0.119, -0.009), 
                           varcovRE =  c(1.000, 0.000, 0.473, 0.000, 1.000, 0.000, 0.057, -0.182,
                                         0.000, 0.174), 
                           fix.varcovRE = c(1, 1, 0, 1, 1, 1, 0, 0, 1, 0),
                           transitionmatrix = c(-0.523, 0.000, 0.000, 0.000, 0.073, 0.083, 
                                                0.119, 0.106, 0.015, 0.157, 0.054, 0.087, 
                                                -0.079, 0.000, 0.000, 0.000), 
                           fix.transitionmatrix = c(0,rep(1,3),rep(0,9),rep(1,3)),
                           var.errors = c(8.876, 0.287, 0.546), 
                           transformationY = c(0.531, 0.256, 1.100,
                                               1.891,  0.846, 1.345))
                                               
mod2 <- DynNet(structural.model = list(fixed.LP0 = ~ 1 + C2 | 1 + C2,
                                     fixed.DeltaLP = L1 + L2| L3  ~ 1 + time| 1 + time,
                                     random.DeltaLP = ~ 1|1,
                                     trans.matrix = ~ 1 + bs(x = time, knots =c(2), 
                                                             intercept = F, degree = 2),
                                     delta.time = Delta),
             measurement.model = list(link.functions = list(links = c(NULL,NULL, NULL),
                                                            knots = list(NULL, NULL, NULL))),
             
             parameters = Parameters,
             option = list(nproc = 2, print.info = TRUE, MCnr = 10, 
                           univarmaxiter = 7, epsa = 1e-5, epsb = 1e-4, epsd = 1e-2),
             Time = "time",
             subject = "id",
             data = data
)

summary(mod2)


      
}
