% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_survival_ui}
\alias{f_survival_ui}
\title{Computes the individual likelihood of time-to-event, conditionally on the random effects}
\usage{
f_survival_ui(
  ui_r,
  t_0i,
  t_i,
  delta_i,
  xti1,
  xti2,
  xti1_intY,
  xti2_intY,
  param_surv,
  param_surv_intY,
  param_basehaz,
  basehaz,
  knots_surv,
  assoc,
  truncation,
  nD,
  DeltaT,
  x0i,
  alpha_mu0,
  xi,
  alpha_mu,
  G_mat_A_0_to_tau_i,
  zi,
  nE
)
}
\arguments{
\item{ui_r}{vector of individual random effects}

\item{t_i}{individual time-to-event}

\item{delta_i}{individual status of event}

\item{xti1}{vector of individual covariates for first event}

\item{xti2}{vector of individual covariates for competing event}

\item{param_surv}{regression parameters}

\item{param_basehaz}{parameters for baseline hazard function}

\item{basehaz}{type of baseline hazard function}

\item{knots_surv}{vector of knots if basehaz == Splines}

\item{assoc}{function of the random effects that captures association 
(0: random intercept, 1: random slope, 2: random intercept and slope, 3: current value, 4: current slope, 5: current value and slope)}

\item{truncation}{boolean, indicating if left truncation or not}
}
\description{
Computes the individual likelihood of time-to-event, conditionally on the random effects
}
