% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fct_pred_curlev_slope}
\alias{fct_pred_curlev_slope}
\title{Computes the hazard risk or survival function for a vector or timepoints}
\usage{
fct_pred_curlev_slope(
  ptGK_delta,
  ptGK,
  xti1,
  xti2,
  xti1_intY,
  xti2_intY,
  ui_r,
  delta_i,
  param_surv,
  param_surv_intY,
  assoc,
  nD,
  DeltaT,
  x0i,
  alpha_mu0,
  xi,
  alpha_mu,
  G_mat_A_0_to_tau_i,
  zi,
  param_basehaz,
  basehaz,
  knots_surv,
  gamma_X,
  nE,
  survfunc,
  trans
)
}
\arguments{
\item{ptGK_delta}{vector of projections of GK nodes onto grid of delta}

\item{ptGK}{vector of individual GK nodes}

\item{delta_i}{event status}

\item{survfunc}{indicator if output is survival function or hazard risk}

\item{trans}{index for computation of survival/risk function on all transitions (-1), on first transition(0), or second transition (1)}

\item{alpha}{vector of association parameters}
}
\description{
Computes the hazard risk or survival function for a vector or timepoints
}
