% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fct_risq_base}
\alias{fct_risq_base}
\title{Computes the hazard risk or survival if surv == true}
\usage{
fct_risq_base(
  t,
  status,
  param_basehaz,
  basehaz,
  knots_surv,
  nE,
  gammaX,
  surv,
  trans
)
}
\arguments{
\item{gammaX}{vector of linear predictors for 1 and 2 transitions (including association on random effects if assoc <=2)}

\item{surv}{Computation of survival funtion (1) or risk function (0)}

\item{trans}{index for computation of survival function on all transitions (-1), on first transition(0), or second transition (1) for nE=2
when}
}
\description{
Computes the hazard risk or survival if surv == true
}
