% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{gapply}
\alias{gapply}
\alias{gapply.dynnet}
\title{Apply Function To Each Network Realization}
\usage{
gapply(network, fn)

\method{gapply}{dynnet}(network, fn)
}
\arguments{
\item{network}{dynnet network object.}

\item{fn}{Function taking an igraph object.}
}
\value{
List of results of applying fn to each realization of the network.
}
\description{
Apply function to each network realization. Currently a wrapper around lapply.
}
\details{
Apply function to each network realization.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

