% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_adjacency}
\alias{generate_adjacency}
\title{Generate Adjacency Matrix from a Set of Latent Positions, Covariate
Profile, and Coefficient Vector}
\usage{
generate_adjacency(pos, Xb, family = NULL, ...)
}
\arguments{
\item{pos}{Matrix of latent positions.}

\item{Xb}{Vector of linear predictors for each dyad.}

\item{family}{String indicating the data generating process to follow.}

\item{...}{Additional parameters to be passed to the distance function.}
}
\value{
An adjacency matrix. These matrices are of a type produced by the
\code{\link{Matrix}} package.
}
\description{
Generate an adjacency matrix from a matrix of latent positions, nodal
covariates, and a vector of coefficients on those covariates.
}
\details{
Generate an adjacency matrix from a matrix of latent positions, nodal
covariates, and a vector of coefficients on those covariates.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

