% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_group}
\alias{generate_group}
\title{Generate a Cluster of Nodes}
\usage{
generate_group(n, mean, sigma)
}
\arguments{
\item{n}{Positive integer. The number of nodes in the cluster.}

\item{mean}{Mean vector. Default is the a zero vector of length \code{k}.}

\item{sigma}{Covariance matrix. Default is a \code{k}-by-\code{k} identity
matrix.}
}
\value{
Numeric matrix with dimensions of \code{(n-1)}-by-\code{k}.
}
\description{
Generate a cluster of nodes in \code{k}-dimensional latent space.
}
\details{
Generate a cluster of nodes in \code{k}-dimensional latent space.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

