% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dynnet.R
\name{get_adjacency}
\alias{get_adjacency}
\title{Extract Adjacency Matrices from dynnet Object}
\usage{
get_adjacency(net, period = NULL)
}
\arguments{
\item{net}{dynnet network.}

\item{period}{Numeric vector.}
}
\value{
List of adjacency matrices, unless a single \code{period} is
specified, then a matrix is returned.
}
\description{
Extract adjacency matrices from dynnet object.
}
\details{
Extract adjacency matrices from dynnet object.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

