% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opt_formative.R
\name{get_opt_formative}
\alias{get_opt_formative}
\title{Recover Lambda-scale parameters and formative weights after optimization}
\usage{
get_opt_formative(b, paras_block_dim, mapping, mapping2, nRE)
}
\arguments{
\item{paras_block_dim}{vector giving the sizes of parameter blocks}

\item{mapping}{vector: outcome → latent process (endogenous)}

\item{mapping2}{vector: outcome → latent process (exogenous)}

\item{temp}{list returned by optimizer (contains temp$b = optimized parameters)}
}
\value{
temp list with parameters re-expressed in Lambda-space
}
\description{
Recover Lambda-scale parameters and formative weights after optimization
}
