% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynnet.R
\name{graphs}
\alias{graphs}
\alias{graphs.dynnet}
\title{Extract Graphs from Network}
\usage{
graphs(network)

\method{graphs}{dynnet}(network)
}
\arguments{
\item{network}{dynnet network object.}
}
\value{
List of graphs representing each period in the network.
}
\description{
The graph object making up the network.
}
\details{
The graph object making up the network.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

