% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/init.R
\name{init_coef}
\alias{init_coef}
\title{Initialize Model Coeffients and Latent Positions}
\usage{
init_coef(Y, X, ref_pos, ref_idx)
}
\arguments{
\item{Y}{Adjacency matrix}

\item{X}{Model matrix for exogenous covariates.}

\item{k}{Positive integer. Number of dimensions for the latent social space.}
}
\value{
Vector of starting values.
}
\description{
Initialize model coeffients and latent positions for latent space and path
models.
}
\details{
Initialize model coeffients and latent positions for latent space and path
models.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

