% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_ref}
\alias{insert_ref}
\title{Insert Reference Locations into Latent Positions Matrix}
\usage{
insert_ref(pos, ref, d)
}
\arguments{
\item{pos}{\code{n-r} by \code{d} positions matrix, where r is the number of}

\item{ref}{Reference object as specified in an latent space model}

\item{d}{Dimensions of the latent space}
}
\value{
Latent positions matrix with reference locations inserted.
}
\description{
Insert the reference vertex locations into the latent positions matrix.
}
\details{
This calls an underlying \code{C} function, \code{.C_insert_ref}, since it's
critical to the speed of the MLE implementation.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
\keyword{internal}

