% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_all_equal}
\alias{list_all_equal}
\title{Check if All List Elements are Equal}
\usage{
list_all_equal(lst)
}
\arguments{
\item{lst}{List.}
}
\value{
Logical indicating whether or not all values of the list are equal.
}
\description{
Checks whether all elements of the supplied list are equal.
}
\details{
Checks whether all elements of the supplied list are equal.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
\keyword{internal}
