% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{lsm}
\alias{lsm}
\title{Static Latent Space Model}
\usage{
lsm(network, k = 1, period = 1, ref = NULL, family = "bernoulli",
  start = start_random, method = "MLE", seed = NULL, verbose = TRUE,
  control = control.lsm())
}
\arguments{
\item{network}{\code{dynnet} network object.}

\item{k}{Integer. Number of dimensions for the latent space.}

\item{period}{Integer. Period to select from the network object.}

\item{ref}{List. Reference vertex information.}

\item{family}{String specifying the model family. Only \code{bernoulli} is
currently supported.}

\item{start}{Function to use to get the starting values for the latent
locations of each vertex.}

\item{method}{String specifying the estimation method. Either \code{MLE}
(maximum likelihood) or \code{MH} (a Metropolis Hastings algorithm).}

\item{seed}{Random seed.}

\item{verbose}{Boolean.}

\item{control}{List of control parameters to be used for}
}
\value{
\code{lsmfit} model object.
}
\description{
Static latent space model for static networks.
}
\details{
Estimate a latent space model for static networks as introduced by HRH
(2002).
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

