% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynnet.R
\name{ndyads.dynnet}
\alias{ndyads.dynnet}
\title{Calculate the Number of Unique Dyads in a Network}
\usage{
ndyads.dynnet(network)
}
\arguments{
\item{network}{A \code{dynnet} network.}
}
\value{
The number of unique dyads in a network.
}
\description{
Calculates the number of unique dyads in a network.
}
\details{
Calculates the number of unique dyads in a network.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
