% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynnet.R
\name{nodes}
\alias{nodes}
\alias{nodes.dynnet}
\title{Number of Nodes in the Network}
\usage{
nodes(network)

\method{nodes}{dynnet}(network)
}
\arguments{
\item{network}{dynnet network object.}
}
\value{
Number of nodes in the (possibly dynamic) network.
}
\description{
The number of nodes in the dynamic network.
}
\details{
The number of nodes in the dynamic network.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

