% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynnet.R
\name{periods.dynnet}
\alias{periods.dynnet}
\title{Number of Periods in the Network}
\usage{
periods.dynnet(net)
}
\arguments{
\item{net}{dynnet network object.}
}
\value{
Number of periods in the (possibly dynamic) network.
}
\description{
The number of periods in the dynamic network.
}
\details{
The number of periods in the dynamic network.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

