% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{positions}
\alias{positions}
\title{Extract Latent Positions from Simulated Network}
\usage{
positions(sim)
}
\arguments{
\item{sim}{Simulated network of class \code{\link{dynnetsim}}.}
}
\value{
Matrix of latent positions.
}
\description{
Extract latent positions from simulated dynamic network.
}
\details{
Extract latent positions from simulated dynamic network.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

