% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print}
\alias{print}
\alias{print.dynnet}
\alias{print.lsmfit}
\title{Print dynnet Network Object}
\usage{
print(object)

\method{print}{dynnet}(object)

\method{print}{lsmfit}(object)
}
\arguments{
\item{network}{dynnet network object.}
}
\value{
Prints.
}
\description{
Print dynnet network object
}
\details{
Print dynnet network object
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
