% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_omega_cov.R
\name{recover_omega_cov}
\alias{recover_omega_cov}
\title{Recover a single Omega-level covariance matrix from a single alternating Lambda-level matrix}
\usage{
recover_omega_cov(Blambda, mappingL1L2)
}
\arguments{
\item{Blambda}{(nD*nTypes) x (nD*nTypes) Lambda-level var/cov (non-Cholesky).
Rows/cols alternate by type: int0, int, slope (or int0, int).}

\item{mappingL1L2}{nD x nL; (d,l) is the weight ω_{d,l} of Omega l in Lambda d.}
}
\value{
A single Omega-level covariance matrix BOmega of size (nL + 2*nL) x (nL + 2*nL)
        ordered as: [all Ω int0] then [Ω1(int), Ω1(slope), Ω2(int), Ω2(slope), ...].
        Attributes include a list of components for inspection.
}
\description{
Recover a single Omega-level covariance matrix from a single alternating Lambda-level matrix
}
