% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{rksphere}
\alias{rksphere}
\title{Random Point From \code{k}-dimensional Unit Hypersphere}
\usage{
rksphere(k)
}
\arguments{
\item{k}{Positive integer indicating the dimensions of the hypersphere.}
}
\value{
Numeric vector of coordinates.
}
\description{
Select a random point from a k-dimensional unit hypersphere.
}
\details{
Select a random point from a k-dimensional unit hypersphere.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

