% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_network}
\alias{simulate_network}
\title{Simulate Dynamic and Static Networks}
\usage{
simulate_network(n, k = 1, periods = 1, groups = 1, X = NULL,
  beta = 0.5, ref_pos = default_ref_pos(k), mean_pos = matrix(0, ncol = k,
  nrow = groups), sigma_pos = rep(list(diag(k)), groups),
  ref_traj = matrix(0, ncol = k, nrow = k + 1), mean_traj = matrix(0, ncol =
  k, nrow = groups), sigma_traj = rep(list(diag(k)), groups),
  traj_fn = trajectory_linear, seed = NULL, family = "logit", ...)
}
\arguments{
\item{n}{Positive integer. Number of nodes in the network.}

\item{k}{Positive integer. Number of dimensions in the latent space.}

\item{periods}{Positive integer. Number of periods in the dynnamic
network. If \code{k=1}, a static network is generated.}

\item{groups}{Positive integer. Number of groups, or clusters, of nodes in
the network.}

\item{X}{List of model matrices of covariates of size \code{N}-by-\code{m},
where \code{N} is the \emph{number of dyads} in the network and \code{m} is
the number of covariates. If \code{NULL} an intercept is used. If a single
matrix is provided, this model matrix will be used for all periods.}

\item{beta}{Vector of coefficients of length \code{m}. Default is
\code{c(0.5)} for the coefficient on the intercept.}

\item{ref_pos}{Matrix representing the positions of the reference units in
the latent space. Should be of dimensions \code{k+1}-by-\code{k}. Default is
to randomly choose \code{k+1} points on a \code{k}-dimensional hypersphere.}

\item{mean_pos}{Matrix representing the mean position of the groups. Should
be of dimensions \code{g}-by-\code{k}. Default is to use the origin for all
groups.}

\item{sigma_pos}{List of \code{g} covariance matrices used to generate the
locations of nodes in each group. Default is a list of \code{g}
\code{k}-by-\code{k} identity matrices.}

\item{ref_traj}{Matrix representing the trajectories of the reference units
through the latent space. Should be of dimensions
\code{k+1}-by-\code{k}. For identification, the locations of the reference
units should ideally be fixed.}

\item{mean_traj}{Matrix representing the mean of the node
trajectories. Default is a \code{k}-by-\code{g} matrix of zeros.}

\item{sigma_traj}{List of \code{g} covariance matrices used to generate the
trajectories. Default is a list of \code{g} \code{k}-by-\code{k} identity
matrices. Note: this may imply too much movement in the nodes.}

\item{traj_fn}{Function taking a positive integer representing the period
for which to generate the network as well as a position and trajectory
matrix. Default is \code{\link{linear_trajectory}}.}

\item{seed}{Set the seed before generating the networks to assure
replicability. Default is \code{NULL}.}

\item{family}{Character string in \code{c("logit", "poisson")}.}

\item{...}{Further parameters to be passed to subsequent functions.}
}
\value{
A \code{dynsim} object.
}
\description{
Simulate dynammic and static networks.
}
\details{
Simulate dynammic and static networks.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

