% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDynNet.R
\name{simulateDynNet}
\alias{simulateDynNet}
\title{Function to simulate data from DynNet model}
\usage{
simulateDynNet(
  Ni,
  structural.model,
  measurement.model,
  parameters,
  TempsFin,
  DeltaT,
  basehaz = NULL,
  assocT = NULL,
  seed = 123,
  varcovRE.format = "cholesky",
  ...
)
}
\arguments{
\item{Ni}{number of individuals}

\item{structural.model}{a list of 7 arguments used to specify the structural model: \describe{
   \item{\code{fixed.LP0}}{a one-sided linear formula object for specifying the 
   fixed effects in the submodel for the baseline level of processes. Note that 
   there is no need to specify a random effect model for the baseline level of 
   processes as we systematically set a process-specific random intercept (with 
   variance fixed to 1 for identifiability purpose). For identifiability purposes, 
   the mean intercepts are fixed to 0 (not estimated).}

   \item{\code{fixed.DeltaLP}}{a two-sided linear formula object for 
   specifying the response outcomes (one the left part of ~ symbol) 
   and the covariates with fixed-effects (on the right part of ~ symbol) 
   in the submodel for change over time of latent processes.}

   \item{\code{random.DeltaLP}}{a one-sided linear formula object 
   for specifying the random effects in the submodel for change 
   over time of latent processes.}

   \item{\code{trans.matrix}}{a one-sided linear formula object 
   for specifying a model for elements of the transition matrix, 
   which captures the temporal influences between latent processes.}

   \item{\code{fixed.survival}}{a one-sided linear formula object 
   for specifying the covariates in the survival sub-model. In competing 
   risks model, the specification for the two events should be separated 
   by the "|" symbol.}

   \item{\code{interactionY.survival}}{a one-sided linear formula 
   object for specifying the covariates in interaction with the 
   dynamics of the latent processes, in the survival sub-model. 
   In competing risks model, the specification for the two events 
   should be separated by the "|" symbol. Only additional terms 
   should be included (No "*" symbol). Covariates in 
   \code{interactionY.survival} should also be included in \code{fixed.survival.}}

   \item{\code{delta.time}}{indicates the discretisation step to be used for latent processes}
   }}

\item{measurement.model}{is a list of arguments detailed below used to specify the measurement model: \describe{

\item{\code{link}}{ indicates the link functions to be used to transform the outcomes. It takes values in "linear" for a linear transformation and "n-type-d" for a I-splines transformation where "n" indicates the number of nodes, "type" (which takes values in "quant", "manual", "equi") indicates where the nodes are placed, and "d" indicates the degree of the I-splines.}

\item{\code{knots}}{ argument indicates if necessary the place of knots (when placed manually with "manual"), default value is NULL}
}}

\item{parameters}{a list of 3 arguments about parameters of the models (e.g., initial parameters, parameters one would like to fix, etc.) obtained with function \link{enter_param} and option \code{full_list=T}: \describe{
\item{\code{paras.ini}}{ indicates initial values for parameters, default values is NULL.} 
\item{\code{Fixed.para.indix}}{ indicates the positions of parameters to be constrained.}

\item{\code{Fixed.para.values}}{ indicates the values associated to the index of parameters to be constrained. }
}}

\item{TempsFin}{Time last observation.}

\item{DeltaT}{Interval between observations.}

\item{basehaz}{baseline hazard for survival model(s) (default NULL).}

\item{assocT}{association structure for survival model(s) (default NULL). Only "value" currently implemented.}

\item{seed}{seed for reproducibility. (default 123)}

\item{varcovRE.format}{character indication format of random effects var/cov matrix parametrization. "cholesky" (default) or "block".}

\item{...}{}
}
\value{
data.frame with simulated data
}
\description{
Function to simulate data from DynNet model
}
