% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate.R
\name{split_n}
\alias{split_n}
\title{Splits \code{n} into \code{g} Groups}
\usage{
split_n(n, g)
}
\arguments{
\item{n}{Integer indicating the total number of nodes in the network.}

\item{g}{Integer indicating the number of groups.}
}
\value{
List of integers of approximately the same size.
}
\description{
Divides an integer into \code{g} approximately equal integers and returns as
a list.
}
\details{
Divides an integer into \code{g} approximately equal integers and returns as
a list.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

