% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynnet.R
\name{to_dynnet}
\alias{to_dynnet}
\alias{to_dynnet.network}
\title{Create dynnet Network Object from Another Format}
\usage{
to_dynnet(network, ...)

\method{to_dynnet}{network}(network, ...)
}
\arguments{
\item{network}{Network object}

\item{...}{Other things}
}
\value{
dynnet network object
}
\description{
Create dynnet Network Object from Another Format
}
\details{
Create dynnet Network Object from Another Format
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
