% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{trajectories}
\alias{trajectories}
\title{Extract Latent Trajectories from Simulated Network}
\usage{
trajectories(sim)
}
\arguments{
\item{sim}{Simulated network of class \code{\link{dynnetsim}}.}
}
\value{
Matrix of latent trajectories.
}
\description{
Extract latent trajectories from simulated dynamic network.
}
\details{
Extract latent trajectories from simulated dynamic network.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

