% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trajectories.R
\name{trajectory_linear}
\alias{trajectory_linear}
\title{Linear Trajectory}
\usage{
trajectory_linear(period, pos, traj, ...)
}
\arguments{
\item{period}{Integer specifying the time period for which to calculate
the latent positions.}

\item{pos}{Matrix of latent locations.}

\item{traj}{Matrix of trajectories.}

\item{...}{Additional parameters to pass. Not used in this trajectory
function.}
}
\value{
Matrix of latent positions.
}
\description{
Simple linear trajectory.
}
\details{
Simple linear trajectory.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

