% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph-utils.R
\name{vcount}
\alias{vcount}
\alias{vcount.igraph}
\alias{vcount.dynnet}
\alias{vcount.lsmfit}
\title{Number of Vertices in a Graph}
\usage{
vcount(object, ...)

\method{vcount}{igraph}(object, ...)

\method{vcount}{dynnet}(object, ...)

\method{vcount}{lsmfit}(object, ...)
}
\arguments{
\item{object}{dynnet network, latent space model fit, or igraph object.}
}
\value{
Number of vertices in the graph.
}
\description{
The number of vertices in a graph
}
\details{
The number of verices in a graph.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}
