% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dynnet.R
\name{dynnet_adjacency}
\alias{dynnet_adjacency}
\title{Dynamic Network from a List of Adjacency Networks}
\usage{
dynnet_adjacency(adj, X = NULL, mode = "lower", weighted = NULL)
}
\arguments{
\item{adj}{List of adjacency matrixes. Single matrices, representing a
static network, are also supported.}

\item{X}{A \code{data.frame} of nodal covariates. Default is \code{NULL}.}

\item{mode}{Type of graph. See \code{\link{igraph::graph.adjacency}}.}

\item{weighted}{Indicates whether the network is weighted
(non-binary). Default is \code{NULL}. See
\code{\link{igraph::graph.adjacency}} for details.}
}
\value{
A \code{dynnet} object.
}
\description{
Generate a dynamic network object from a list of adjacency matrices.
}
\details{
Generate a dynamic network object from a list of adjacency matrices.
}
\author{
Jason W. Morgan \email{jason.w.morgan@gmail.com}
}

