% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{locations}
\alias{locations}
\alias{locations.lsmfit}
\title{Estimated Locations from Latent Space Model}
\usage{
locations(object, ...)

\method{locations}{lsmfit}(object, transform = "procrustes", ...)
}
\arguments{
\item{object}{A \code{\link{lsmfit}} object.}

\item{...}{Addition parameters to pass.}
}
\value{
A matrix of estimated locations. See Details.
}
\description{
Extract the estimated latent locations for a latent space model.
}
\details{
This function returns the estimated latent locations for a latent space
model. For a model fit via MLE, this is the MLE estimates. For models with
with the Metropolis-Hastings algorithm, the posterior means for the
locations (possibly transformed via a Procrustes step, if that was the
method of identification specified). Note: the posterior means should not be
used for inference as they can distort the appearance of vertices in the the
latent space. Use the posterior samples instead.
}
\author{
Jason Morgan
}
